+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "open-import-file" extension will be documented in th
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+### Changed
+
+- Update repository location
+
 ## [0.1.1] - 2025-10-16
 
 - Fixed resolving file paths starting with `@/` did not work correctly.